refactor: make string expression to class
